Move input onChange functions from StudentDrawerFilter to their specific Input component. Instead, these Input components will accept a setFilter and/or deleteFilter function to keep the component flexible.
