+++ pharmd-app/src/components/Inputs/GpaSliderInput.js
@@ -3,7 +3,6 @@ import { useInput } from "react-admin";
 import RangeSlider from "../Basic/RangeSlider";
 
 function valuetext(value) {
-  console.log("VALUES", value);
   return `${value}`;
 }
 
@@ -12,9 +11,18 @@ const GpaSliderInput = props => {
     meta: { error }
   } = useInput(props);
 
+  const searchGpaRange = (event, newValue) => {
+    if (newValue) {
+      let gpaMin = newValue[0];
+      let gpaMax = newValue[1];
+      props.setFilter("gpa_gte", gpaMin);
+      props.setFilter("gpa_lte", gpaMax);
+    }
+  };
+
   return (
     <RangeSlider
-      onChange={props.onChange}
+      onChange={searchGpaRange}
       setValueText={valuetext}
       disabled={error}
       max={4}
