+++ pharmd-app/src/screens/student/StudentDrawerFilter.js
@@ -53,7 +53,6 @@ const Filter = styled(FilterRA)`
 ` ;
 
 export const StudentDrawerFilter = props => {
-  console.log("FILTER DRAWER PROPS", props.filterValues);
 
   const classes = useStyles();
 
@@ -67,80 +66,42 @@ export const StudentDrawerFilter = props => {
     props.setFilters(props.filterValues);
   };
 
-  const searchGpaRange = (event, newValue) => {
-    if (newValue) {
-      let gpaMin = newValue[0];
-      let gpaMax = newValue[1];
-      setFilter("gpa_gte", gpaMin);
-      setFilter("gpa_lte", gpaMax);
-    }
-  };
-
-  const addOriginFilter = (newValue) => {
-    switch (newValue) {
-      case "domestic":
-        setFilter("international", false);
-        break;
-      case "international":
-        setFilter("international", true);
-        break;
-      default:
-        deleteFilter("international");
-        break;
-    }
-    console.log("Origin Filter:", newValue);
-  };
-
-  const originCheckbox = (event, array) => {
-    array.length === 1 ?
-      addOriginFilter(array[0])
-      : deleteFilter("international");
-  };
-
-  const statusCheckbox = (event, array) => {
-    setFilter("status", array);
-    console.log("Status Filter:", array);
-  };
-
-  const cohortMultiSelect = (event, array) => {
-    setFilter("cohort[current]", array);
-  };
-
   return (
     <Filter {...props}>
-      {/* <NumberInput label="Min GPA" source={"gpa_gte"} alwaysOn /> */}
-      {/* <NumberInput label="Max GPA" source={"gpa_lte"} alwaysOn /> */}
-
       <StatusCheckboxInput
         label="Status"
         source={"status"}
-        onChange={statusCheckbox}
         color="primary"
         className={classes.formControl}
         checkboxClassName={classes.checkboxButton}
+        setFilter={setFilter}
+        deleteFilter={deleteFilter}
         alwaysOn
       />
       <GpaSliderInput
         label="GPA"
         source={"gpa_gte"}
-        onChange={searchGpaRange}
         className={classes.rangeSlider}
+        setFilter={setFilter}
+        deleteFilter={deleteFilter}
         alwaysOn
       />
       <CohortMultipleSelect
         label="Cohort"
         source={"cohort"}
-        onChange={cohortMultiSelect}
         className={classes.formControl}
+        setFilter={setFilter}
+        deleteFilter={deleteFilter}
         alwaysOn
       />
       <OriginCheckboxInput
         label="Origin"
         source={"international"}
-        onChange={originCheckbox}
         color="primary"
         className={classes.formControl}
         checkboxClassName={classes.checkboxButton}
+        setFilter={setFilter}
+        deleteFilter={deleteFilter}
         alwaysOn
       />
     </Filter>
